+++ src/main.ts
@@ -5,6 +5,7 @@ import router from './router'
 import { createHead } from '@vueuse/head'
 import { store } from './store'
 import './index.css'
+import * as volar from '@volar/experimental/client'
 
 const head = createHead()
 const app = createApp(App)
@@ -14,5 +15,6 @@ app.component('BaseTemplate', BaseTemplate)
 app.use(store)
 app.use(router)
 app.use(head)
+app.use(volar.vuePlugin)
 
 app.mount('#app')